[Chore] Separate out attention backend constants from vllm.utils
Part of https://github.com/vllm-project/vllm/issues/26900

This PR extracts attention backend-related constants from the monolithic
vllm/utils/__init__.py into a dedicated vllm/attention/constants.py module.

Changes:
- Created vllm/attention/constants.py with all attention backend constants
- Updated vllm/utils/__init__.py with deprecation warnings for backward compatibility
- Updated imports in vllm/attention/selector.py to use the new module
- Updated test files to import from the new location:
  - tests/kernels/attention/test_attention_selector.py
  - tests/kernels/attention/test_rocm_attention_selector.py
  - tests/models/quantization/test_fp8.py

The following constants have been moved:
- vllm.utils.STR_BACKEND_ENV_VAR -> vllm.attention.constants.STR_BACKEND_ENV_VAR
- vllm.utils.STR_FLASHINFER_ATTN_VAL -> vllm.attention.constants.STR_FLASHINFER_ATTN_VAL
- vllm.utils.STR_TORCH_SDPA_ATTN_VAL -> vllm.attention.constants.STR_TORCH_SDPA_ATTN_VAL
- vllm.utils.STR_XFORMERS_ATTN_VAL -> vllm.attention.constants.STR_XFORMERS_ATTN_VAL
- vllm.utils.STR_FLASH_ATTN_VAL -> vllm.attention.constants.STR_FLASH_ATTN_VAL
- vllm.utils.STR_INVALID_VAL -> vllm.attention.constants.STR_INVALID_VAL

Backward compatibility is maintained through deprecation warnings.
