+++ tests/kernels/attention/test_rocm_attention_selector.py
@@ -4,9 +4,9 @@
 import pytest
 import torch
 
+from vllm.attention.constants import STR_BACKEND_ENV_VAR
 from vllm.attention.selector import _cached_get_attn_backend, get_attn_backend
 from vllm.platforms.rocm import RocmPlatform
-from vllm.utils import STR_BACKEND_ENV_VAR
 
 
 @pytest.fixture(autouse=True)
